requirements.txt
Django==1.8.5
django-ajax-selects==1.3.6
django-autocomplete-light==2.2.10
django-storages==1.1.8
gunicorn==19.3.0
south==1.0.2
boto==2.38.0
psycopg2==2.6.1